Fix mistakes

+++ LabN2.cpp
@@ -4,9 +4,9 @@ using namespace std;
 
 int main()
 {
-	int y;
+	long double y;
 	cin >> y;
-	long float result = (7 * t(0.25f) + 2 * t(1 + y)) / (6 - t(pow(y, 2) - 1));
+	long double result = (7 * t(0.25f) + 2 * t(1 + y)) / (6 - t(pow(y, 2) - 1));
 	if (isinf(result))
 	{
 		cout << "Result = " << 0;
